BackEnd/requirements.txt
asgiref==3.8.1
boto3==1.38.4
botocore==1.38.4
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
Django==5.1.7
django-cors-headers==4.7.0
django-storages==1.14.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
Farama-Notifications==0.0.4
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.1.1
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
mysqlclient==2.2.7
networkx==3.4.2
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pathfinding==1.0.16
pillow==11.1.0
pygame==2.6.1
PyJWT==2.9.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
s3transfer==0.12.0
setuptools==76.0.0
Shimmy==2.0.0
six==1.17.0
sqlparse==0.5.3
stable_baselines3==2.5.0
swig==4.3.0
sympy==1.13.1
torch==2.6.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.4.0
